Fix upload bug
