+++ samples/apps/copilot-chat-app/webapp/src/components/chat/chat-list/NewBotMenu.tsx
@@ -1,45 +1,29 @@
 // Copyright (c) Microsoft. All rights reserved.
 
-import { FC, useCallback, useRef, useState } from 'react';
+import { FC, useState } from 'react';
 
 import { Button, Menu, MenuItem, MenuList, MenuPopover, MenuTrigger, Tooltip } from '@fluentui/react-components';
 import { ArrowUploadRegular, BotAdd20Filled, BotAdd20Regular, PeopleTeamAddRegular, bundleIcon } from '@fluentui/react-icons';
-import { AlertType } from '../../../libs/models/AlertType';
-import { Bot } from '../../../libs/models/Bot';
 import { useChat } from '../../../libs/useChat';
-import { useFile } from '../../../libs/useFile';
-import { useAppDispatch } from '../../../redux/app/hooks';
-import { addAlert } from '../../../redux/features/app/appSlice';
-import { FileUploader } from '../../FileUploader';
 import { InvitationJoinDialog } from '../invitation-dialog/InvitationJoinDialog';
 
-export const NewBotMenu: FC = () => {
-    const dispatch = useAppDispatch();
+interface NewBotMenuProps {
+    onFileUpload: () => void;
+}
+
+export const NewBotMenu: FC<NewBotMenuProps> = ({
+    onFileUpload
+}) => {
     const chat = useChat();
-    const fileHandler = useFile();
     // It needs to keep the menu open to keep the FileUploader reference
     // when the file uploader is clicked.
     const [isJoiningBot, setIsJoiningBot] = useState(false);
 
-    const fileUploaderRef = useRef<HTMLInputElement>(null);
     const BotAdd20 = bundleIcon(BotAdd20Filled, BotAdd20Regular)
 
     const onAddChat = () => {
         chat.createChat();
     };
-
-    const onUpload = useCallback(
-        (file: File) => {
-            fileHandler
-                .loadFile<Bot>(file, chat.uploadBot)
-                .catch((error) =>
-                    dispatch(
-                        addAlert({ message: `Failed to parse uploaded file. ${error.message}`, type: AlertType.Error }),
-                    ),
-                );
-        },
-        [fileHandler, chat, dispatch],
-    );
 
     const onJoinClick = () => {
         setIsJoiningBot(true);
@@ -67,14 +51,9 @@ export const NewBotMenu: FC = () => {
                         </MenuItem>
                         <MenuItem
                             icon={<ArrowUploadRegular />}
-                            onClick={() => fileUploaderRef.current?.click()}
+                            onClick={onFileUpload}
                         >
                             <div>Upload a Bot</div>
-                            <FileUploader
-                                ref={fileUploaderRef}
-                                acceptedExtensions={['.txt', '.json']}
-                                onSelectedFile={onUpload}
-                            />
                         </MenuItem>
                         <MenuItem icon={<PeopleTeamAddRegular />} onClick={onJoinClick}>
                             Join a Bot
